stored the key in separate file
