+++ libs/query/src/lib/aa/trigger-statements/index.ts
@@ -1,2 +1,3 @@
 export * from './trigger-statements.service';
 export * from './trigger-statements.store';
+export * from './trigger-statements.constants';
